requirements.txt
streamlit
pandas
pillow
PyPDF2
matplotlib
openpyxl